Separate components of the abstract graph experiment into headers
