+++ src/abstractGraph/abstractGraph.cpp
@@ -10,169 +10,10 @@
 #include <typeindex>
 #include <vector>
 
-struct FrameBase {
-    FrameBase() = default;
-};
+#include "FrameBase.h"
+#include "NodeBase.h"
+#include "CustomTypes.h"
 
-template <typename T>
-concept Frame = 
-    std::is_base_of_v<FrameBase, T> 
-    && requires(T a, const T b, const T& c) {
-        T();
-        { a.clone() } -> std::same_as<T>;
-        { b + c } -> std::same_as<T>;
-        { a += b } -> std::same_as<T&>;
-        { a.reset() } -> std::same_as<void>;
-        
-        // { b * c } -> std::same_as<T>;
-        // { a *= b } -> std::same_as<void>;
-};
-
-struct AbstractNode {
-    std::string name;
-    AbstractNode(const char *name) : name(std::string(name)) { }
-    virtual std::type_index getTypeIdInput() = 0;
-    virtual std::type_index getTypeIdOutput() = 0;
-    virtual void reset() = 0;
-};
-
-template <Frame InputT>
-struct NodeInput {
-    virtual void processNext(const InputT& input) = 0;
-};
-
-template <Frame OutputT>
-struct NodeOutput {
-    virtual OutputT getResult() = 0;
-};
-
-template <Frame InputT, Frame OutputT>
-struct Node : public NodeInput<InputT>, public NodeOutput<OutputT>, public AbstractNode {
-    Node(const char *nodeName) : AbstractNode(nodeName), lastOutput({}) { }
-
-    void processNext(const InputT& input) override {
-        lastOutput = tick(input);
-    }
-
-    virtual OutputT tick(InputT input) = 0;
-
-    OutputT getResult() override {
-        return lastOutput.clone();
-    }
-
-    std::type_index getTypeIdInput() override {
-        return typeid(InputT);
-    }
-
-    std::type_index getTypeIdOutput() override {
-        return typeid(OutputT);
-    }
-
-    void reset() override {
-        lastOutput.reset();
-    }
-    
-    NodeInput<InputT>* asInput() {
-        return dynamic_cast<NodeInput<InputT>*>(this);
-    }
-
-    NodeOutput<OutputT>* asOutput() {
-        return dynamic_cast<NodeOutput<OutputT>*>(this);
-    }
-    
-protected: 
-    OutputT lastOutput;
-};
-
-////////
-// example implementation
-
-struct NullFrame : public FrameBase {
-    NullFrame() {}
-    NullFrame& operator+=(const NullFrame &other) {
-        return *(this);
-    }
-    NullFrame operator+(const NullFrame &other) const {
-        return NullFrame();
-    }
-    NullFrame clone() {
-        return NullFrame();
-    }
-    void reset() {}
-};
-
-struct IntFrame : public FrameBase {
-    IntFrame() : data(0) {}
-    int data;
-
-    IntFrame& operator+=(const IntFrame &other) {
-        data += other.data;
-        return *(this);
-    }
-    IntFrame operator+(const IntFrame &other) const {
-        IntFrame f;
-        f.data = data + other.data;
-        return f;
-    }
-    IntFrame clone() {
-        IntFrame f;
-        f.data = data;
-        return f;
-    }
-    void reset() {
-        data = 0;
-    }
-};
-
-struct FloatFrame : public FrameBase {
-    FloatFrame() : data(0.f) {}
-    float data;
-
-    FloatFrame& operator+=(const FloatFrame &other) {
-        data += other.data;
-        return *(this);
-    }
-    FloatFrame operator+(const FloatFrame &other) const {
-        FloatFrame f;
-        f.data = data + other.data;
-        return f;
-    }
-    FloatFrame clone() {
-        FloatFrame f;
-        f.data = data;;
-        return f;
-    }
-    void reset() {
-        data = 0.f;
-    }
-};
-
-struct SourceNode : public Node<NullFrame, IntFrame> {
-    using Node<NullFrame, IntFrame>::Node;
-    IntFrame tick(NullFrame input) {
-        IntFrame f;
-        f.data = 1;
-        return f;
-    }
-};
-
-struct UpcastNode : public Node<IntFrame, FloatFrame> {
-    using Node<IntFrame, FloatFrame>::Node;
-    FloatFrame tick(IntFrame input) {
-        FloatFrame f;
-        f.data = static_cast<float>(input.data);
-        return f;
-    }
-};
-
-struct FloatNode : public Node<FloatFrame, FloatFrame> {
-    using Node<FloatFrame, FloatFrame>::Node;
-    FloatFrame tick(FloatFrame input) {
-        FloatFrame f;
-        f.data = input.data + lastOutput.data;
-        return f;
-    }
-};
 
 
 template<Frame FrameType>
